refactor: DTO를 표현계층으로 분리
