--- src/main/java/es/princip/ringus/application/auth/dto/response/LoginResponse.java
@@ -1,5 +1,0 @@
-package es.princip.ringus.application.auth.dto.response;
-
-public record LoginResponse(
-    Long id
-){ }
