--- src/main/java/es/princip/ringus/application/auth/dto/response/SignUpResponse.java
@@ -1,6 +1,0 @@
-package es.princip.ringus.application.auth.dto.response;
-
-public record SignUpResponse(
-    Long id
-) {
-}
