+++ src/main/java/es/princip/ringus/application/member/service/MemberService.java
@@ -1,6 +1,6 @@
 package es.princip.ringus.application.member.service;
 
-import es.princip.ringus.application.auth.dto.request.SignUpRequest;
+import es.princip.ringus.presentation.auth.dto.request.SignUpRequest;
 import es.princip.ringus.domain.exception.SignUpErrorCode;
 import es.princip.ringus.domain.member.Member;
 import es.princip.ringus.domain.member.MemberRepository;
