+++ src/main/java/es/princip/ringus/presentation/auth/AuthController.java
@@ -1,11 +1,11 @@
-package es.princip.ringus.presentation;
+package es.princip.ringus.presentation.auth;
 
-import es.princip.ringus.application.auth.dto.request.GenerateCodeRequest;
-import es.princip.ringus.application.auth.dto.request.LoginRequest;
-import es.princip.ringus.application.auth.dto.request.SignUpRequest;
-import es.princip.ringus.application.auth.dto.request.VerifyRequest;
-import es.princip.ringus.application.auth.dto.response.LoginResponse;
-import es.princip.ringus.application.auth.dto.response.SignUpResponse;
+import es.princip.ringus.presentation.auth.dto.request.GenerateCodeRequest;
+import es.princip.ringus.presentation.auth.dto.request.LoginRequest;
+import es.princip.ringus.presentation.auth.dto.request.SignUpRequest;
+import es.princip.ringus.presentation.auth.dto.request.VerifyRequest;
+import es.princip.ringus.presentation.auth.dto.response.LoginResponse;
+import es.princip.ringus.presentation.auth.dto.response.SignUpResponse;
 import es.princip.ringus.application.auth.service.AuthService;
 import es.princip.ringus.application.auth.service.EmailVerificationService;
 import es.princip.ringus.global.util.ApiResponse;
