+++ src/main/java/es/princip/ringus/presentation/auth/dto/response/LoginResponse.java
@@ -1,0 +1,5 @@
+package es.princip.ringus.presentation.auth.dto.response;
+
+public record LoginResponse(
+    Long id
+){ }
