+++ src/main/java/es/princip/ringus/presentation/auth/dto/response/SignUpResponse.java
@@ -1,0 +1,6 @@
+package es.princip.ringus.presentation.auth.dto.response;
+
+public record SignUpResponse(
+    Long id
+) {
+}
